Refactor System::myCycles to be 64-bit instead of 32-bit. - The cycles counter is now essentially monotonically increasing (ie, we never need to worry about it going backwards and giving a negative difference, simplifying a lot of code - There are now reset() methods in all places that keep track of system cycles, but they are used for a full reset only; not called each frame like before (which had to be done to prevent overflow).

+++ src/emucore/CartWD.cxx
@@ -275,7 +275,7 @@ bool CartridgeWD::save(Serializer& out) const
     out.putString(name());
     out.putShort(myCurrentBank);
     out.putByteArray(myRAM, 64);
-    out.putInt(myCyclesAtBankswitchInit);
+    out.putLong(myCyclesAtBankswitchInit);
     out.putShort(myPendingBank);
   }
   catch(...)
@@ -297,7 +297,7 @@ bool CartridgeWD::load(Serializer& in)
 
     myCurrentBank = in.getShort();
     in.getByteArray(myRAM, 64);
-    myCyclesAtBankswitchInit = in.getInt();
+    myCyclesAtBankswitchInit = in.getLong();
     myPendingBank = in.getShort();
 
     bank(myCurrentBank);
